transform-load-weather-data/requirements.txt
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0